wordpress nonce fields

+++ plugins/lwn-options-general/lwn-options-general.php
@@ -43,6 +43,19 @@ add_action(
 );
 function lwn_options_general_process_options_callback()
 {
+  $nonce_value = sanitize_text_field($_POST['lwn_options_manager_nonce'] ?? '');
+  if (!wp_verify_nonce($nonce_value, 'lwn_options_manager_nonce')) {
+    wp_die('Invalid nonce provided', 'Error', ['response' => 403]);
+  }
+  // Check Nonce
+  if (
+    !check_admin_referer(
+      'lwn_options_manager_nonce',
+      'lwn_options_manager_nonce',
+    )
+  ) {
+    wp_die('Invalid nonce provided', 'Error', ['response' => 403]);
+  }
   // Process Form
   /* $option_name = isset($_POST['lwn_option_name']) */
   /*   ? $_POST['lwn_option_name'] */
@@ -54,17 +67,31 @@ function lwn_options_general_process_options_callback()
 
   // Process Actions
   if ($action === 'add' || $action === 'update') {
-    update_option($option_name, $option_value);
-    $output = "<div class='updated'>";
-    $output .= __('Option saved successfully', 'text-domain');
-    $output .= '</div>';
-    echo $output;
+    if (empty($option_name) || empty($option_value)) {
+      $output = "<div class='error'><p>";
+      $output .= __('Option Name and option value are required', 'text-domain');
+      $output .= '</p></div>';
+      echo $output;
+    } else {
+      update_option($option_name, $option_value);
+      $output = "<div class='updated'>";
+      $output .= __('Option saved successfully', 'text-domain');
+      $output .= '</div>';
+      echo $output;
+    }
   } elseif ($action === 'delete') {
-    delete_option($option_name);
-    $output = "<div class='updated'>";
-    $output .= __('Option deleted successfully', 'text-domain');
-    $output .= '</div>';
-    echo $output;
+    if (empty($option_name)) {
+      $output = "<div class='error'><p>";
+      $output .= __('Option Name is required', 'text-domain');
+      $output .= '</p></div>';
+      echo $output;
+    } else {
+      delete_option($option_name);
+      $output = "<div class='updated'>";
+      $output .= __('Option deleted successfully', 'text-domain');
+      $output .= '</div>';
+      echo $output;
+    }
   } else {
     $output = "<div class='error'>";
     $output .= __('Invalid action', 'text-domain');
@@ -77,10 +104,20 @@ function lwn_options_general_process_options_callback()
 add_action('lwn_options_general_form', 'lwn_options_general_form_callback');
 function lwn_options_general_form_callback()
 {
+  $option_name = isset($_POST['lwn_option_name'])
+    ? $_POST['lwn_option_name']
+    : '';
+  $option_value = isset($_POST['lwn_option_value'])
+    ? $_POST['lwn_option_value']
+    : '';
   ?>
 			<div class="wrap">
 				<h1><?php echo __('Options Manager', 'text-domain'); ?></h1>
 				<form method="post">
+            <?php wp_nonce_field(
+              'lwn_options_manager_nonce',
+              'lwn_options_manager_nonce',
+            ); ?>
 						<table class="form-table">
                 <tr>
 										<th class="row">
@@ -89,7 +126,9 @@ function lwn_options_general_form_callback()
 											</label>
                      </th>
 										 <td>
-                          <input type="text" id="lwn_option_name" name="lwn_option_name" class="regular-text" required>
+                     <input type="text" value="<?php echo esc_attr(
+                       $option_name,
+                     ); ?>" id="lwn_option_name" name="lwn_option_name" class="regular-text" >
                      </td>
 								</tr>
                 <tr>
@@ -99,7 +138,9 @@ function lwn_options_general_form_callback()
 											</label>
                      </th>
 										 <td>
-                          <input type="text" id="lwn_option_value" name="lwn_option_value" class="regular-text" required>
+                     <input value="<?php echo esc_attr(
+                       $option_value,
+                     ); ?>" type="text" id="lwn_option_value" name="lwn_option_value" class="regular-text" >
                      </td>
 								</tr>
                 <tr>
